databricks-builder-app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".", "packages"]
include = ["server*", "databricks_tools_core*", "databricks_mcp_server*"]
exclude = ["client*", "tests*", "scripts*"]

[project]
name = "databricks-builder-app"
version = "0.1.0"
description = "Claude Code MCP Application"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "fastapi[standard]>=0.115.8",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.34.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "databricks-sdk>=0.81.0",
  # Database
  "sqlalchemy[asyncio]>=2.0.41",
  "alembic>=1.16.1",
  "psycopg[binary]>=3.2.0",  # psycopg3 async driver with hostaddr support
  "greenlet>=3.0.0",
  "psycopg2-binary>=2.9.11",  # For alembic migrations (sync)
  # Claude Agent SDK (successor to claude-code-sdk)
  "claude-agent-sdk>=0.1.55",
  "anthropic>=0.42.0",
  # Databricks MCP tools (databricks-mcp-server installed from sibling dir in dev)
  "mcp>=1.0.0",
  "fastmcp>=0.1.0",
  # databricks-tools-core and databricks-mcp-server are bundled in packages/ directory
  "requests>=2.31.0",
  "sqlglot>=20.0.0",
  "sqlfluff>=3.0.0",
  "litellm>=1.0.0",
  "pymupdf>=1.24.0",
  # Conflict resolution pins for Databricks Apps pre-installed packages
  "tenacity==9.0.0",
  "pillow==11.1.0",
  "websockets==15.0",
  "pyarrow==18.1.0",
  "markupsafe==3.0.2",
  "rich==14.0.0",
  "protobuf>=3.12.0,<5",
  "pandas==2.3.0",
  "flask==3.1.0",
  "werkzeug==3.1.3",
]

[dependency-groups]
dev = [
  "click>=8.1.8",
  "ruff>=0.9.6",
  "watchdog[watchmedo]>=6.0.0",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
src = ["server"]
line-length = 100
indent-width = 2
target-version = "py311"

exclude = [
  "__pycache__",
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "W", "F", "Q", "I", "D"]
ignore = [
  "D105",
  "D106",
  "D107",
  "D203",
  "D204",
  "D213",
  "D401",
  "D407",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"*_test.py" = ["D101", "D102", "D103", "D104"]
"__init__.py" = ["D104"]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "pydantic.validator"]

[tool.ruff.lint.pydocstyle]
convention = "google"
